requirements.txt
<<<<<<< HEAD
pyTelegramBotAPI==4.22.1
google-generativeai==0.8.3
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
=======
google-generativeai==0.8.3
python-telegram-bot>=21.0
gunicorn==22.0.0
uvicorn==0.30.6
fastapi==0.115.0
>>>>>>> 20530eabdcfa35475f9cac732925d92320869426
